requirements.txt
discord.py
# sqlite3